feat: remove kadane and replace it with kadane2
